# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "API for an employee management system to determine employees that\
    \ are no longer working on active projects and to understand their qualifications\
    \ and expertise for suitability in other projects"
  version: "1.0.0"
  title: "UBahn API"
...
schemes:
- "https"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /skills:
    get:
      tags:
      - "Skills"
      description: "Get list of skills in the application.  If no results, then empty\
        \ array is returned.  Multiple filters are\nsupported.\n\n**Security** - Note\
        \ that for non-admin users, this endpoint will only return entities that\n\
        the user has created.\n"
      operationId: "skillsGET"
      parameters: []
      responses:
        "200":
          description: "OK - the request was successful"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Skill"
        "400":
          $ref: "#/definitions/BadRequest"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "Skills"
    head:
      tags:
      - "Skills"
      description: "Retrieve header information for get operation on Skills in the\
        \ application.\n\n**Security** - Note that for non-admin users, this endpoint\
        \ will only return entities that\nthe user has created.\n"
      operationId: "skillsHEAD"
      parameters: []
      responses:
        "200":
          description: "Success response"
        "400":
          $ref: "#/definitions/BadRequest"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "Skills"
    post:
      tags:
      - "Skills"
      description: "Create a new Skill.\n\n**Security** - This endpoint is accessible\
        \ by all authenticated users.        \n"
      operationId: "skillsPOST"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/SkillRequestBody"
      responses:
        "200":
          description: "OK - the request was successful"
          schema:
            $ref: "#/definitions/Skill"
        "400":
          $ref: "#/definitions/BadRequest"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "409":
          $ref: "#/definitions/Conflict"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "Skills"
  /skills/{skillId}:
    get:
      tags:
      - "Skills"
      description: "Get Skill by given skill id.\n\n**Security** - Note that for non-admin\
        \ users, this endpoint will only return entities that\nthe user has created.\
        \        \n"
      operationId: "skillsSkillIdGET"
      parameters:
      - name: "skillId"
        in: "path"
        description: "The skill id"
        required: true
        type: "string"
        format: "UUID"
      responses:
        "200":
          description: "OK - the request was successful"
          schema:
            $ref: "#/definitions/Skill"
        "400":
          $ref: "#/definitions/BadRequest"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "404":
          $ref: "#/definitions/NotFound"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "Skills"
    head:
      tags:
      - "Skills"
      description: "Retrieve header information for get operation on Skill by its\
        \ Id in the application.\n\n**Security** - Note that for non-admin users,\
        \ this endpoint will only return entities that\nthe user has created.    \
        \   \n"
      operationId: "skillsSkillIdHEAD"
      parameters:
      - name: "skillId"
        in: "path"
        description: "The user id"
        required: true
        type: "string"
        format: "UUID"
      responses:
        "200":
          description: "OK - the request was successful"
        "400":
          $ref: "#/definitions/BadRequest"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "404":
          $ref: "#/definitions/NotFound"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "Skills"
    delete:
      tags:
      - "Skills"
      description: "Remove an existing skill with given id.\n\n**Security** - Note\
        \ that this endpoint is only available for admin users.  \n"
      operationId: "skillsSkillIdDELETE"
      parameters:
      - name: "skillId"
        in: "path"
        description: "The skill id"
        required: true
        type: "string"
        format: "UUID"
      responses:
        "204":
          description: "OK - the request was successful"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "404":
          $ref: "#/definitions/NotFound"
        "409":
          $ref: "#/definitions/Conflict"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "Skills"
    patch:
      tags:
      - "Skills"
      description: "Update an existing skill with given id.\n\n**Security** - Note\
        \ that for non-admin users, this endpoint will only allow updates on entities\
        \ that the\ncalling user has created.\n"
      operationId: "skillsSkillIdPATCH"
      parameters:
      - name: "skillId"
        in: "path"
        description: "The skill id"
        required: true
        type: "string"
        format: "UUID"
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/SkillUpdateRequestBody"
      responses:
        "200":
          description: "OK - the request was successful"
          schema:
            $ref: "#/definitions/Skill"
        "400":
          $ref: "#/definitions/BadRequest"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "404":
          $ref: "#/definitions/NotFound"
        "409":
          $ref: "#/definitions/Conflict"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "Skills"
...
  /roles:
    get:
      tags:
      - "Roles"
      description: "Search Roles in the application.  If no results, then empty array\
        \ is returned.\n\n**Security** - Note that for non-admin users, this endpoint\
        \ will only return entities that\nthe user has created.\n"
      operationId: "rolesGET"
      parameters:
      - name: "name"
        in: "query"
        description: "Filter by role name"
        required: false
        type: "string"
      responses:
        "200":
          description: "OK - the request was successful"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Role"
        "400":
          $ref: "#/definitions/BadRequest"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "Roles"
    head:
      tags:
      - "Roles"
      description: "Retrieve header information for a search operation on Roles in\
        \ the application.\n\n**Security** - Note that for non-admin users, this endpoint\
        \ will only return entities that\nthe user has created.\n"
      operationId: "rolesHEAD"
      parameters:
      - name: "name"
        in: "query"
        description: "Filter by role name"
        required: false
        type: "string"
      responses:
        "200":
          description: "Success response"
        "400":
          $ref: "#/definitions/BadRequest"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "Roles"
    post:
      tags:
      - "Roles"
      description: "Create a new Role.\n\n**Security** - This endpoint is accessible\
        \ by all authenticated users.        \n"
      operationId: "rolesPOST"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/NameRequestBody"
      responses:
        "200":
          description: "OK - the request was successful"
          schema:
            $ref: "#/definitions/Role"
        "400":
          $ref: "#/definitions/BadRequest"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "409":
          $ref: "#/definitions/Conflict"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "Roles"
  /roles/{roleId}:
    get:
      tags:
      - "Roles"
      description: "Get role with given id.\n\n**Security** - Note that for non-admin\
        \ users, this endpoint will only return entities that\nthe user has created.\
        \        \n"
      operationId: "rolesRoleIdGET"
      parameters:
      - name: "roleId"
        in: "path"
        description: "The role id"
        required: true
        type: "string"
        format: "UUID"
      responses:
        "200":
          description: "OK - the request was successful"
          schema:
            $ref: "#/definitions/Role"
        "400":
          $ref: "#/definitions/BadRequest"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "404":
          $ref: "#/definitions/NotFound"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "Roles"
    head:
      tags:
      - "Roles"
      description: "Get role with given id, but only header information is returned.\n\
        \n**Security** - Note that for non-admin users, this endpoint will only return\
        \ entities that\nthe user has created.        \n"
      operationId: "rolesRoleIdHEAD"
      parameters:
      - name: "roleId"
        in: "path"
        description: "The role id"
        required: true
        type: "string"
        format: "UUID"
      responses:
        "200":
          description: "OK - the request was successful"
        "400":
          $ref: "#/definitions/BadRequest"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "404":
          $ref: "#/definitions/NotFound"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "Roles"
    delete:
      tags:
      - "Roles"
      description: "Remove an existing role with given id.\n\n**Security** - Note\
        \ that this endpoint is only available for admin users.  \n"
      operationId: "rolesRoleIdDELETE"
      parameters:
      - name: "roleId"
        in: "path"
        description: "The role id"
        required: true
        type: "string"
        format: "UUID"
      responses:
        "204":
          description: "OK - the request was successful"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "404":
          $ref: "#/definitions/NotFound"
        "409":
          $ref: "#/definitions/Conflict"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "Roles"
    patch:
      tags:
      - "Roles"
      description: "Update an existing role with given id.\n\n**Security** - Note\
        \ that for non-admin users, this endpoint will only allow updates on entities\
        \ that the\ncalling user has created.\n"
      operationId: "rolesRoleIdPATCH"
      parameters:
      - name: "roleId"
        in: "path"
        description: "The role id"
        required: true
        type: "string"
        format: "UUID"
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/NameRequestBody"
      responses:
        "200":
          description: "OK - the request was successful"
          schema:
            $ref: "#/definitions/Role"
        "400":
          $ref: "#/definitions/BadRequest"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "404":
          $ref: "#/definitions/NotFound"
        "409":
          $ref: "#/definitions/Conflict"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "Roles"
...
securityDefinitions:
  Bearer:
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
  AuditFields:
    type: "object"
    required:
    - "created"
    - "createdBy"
    - "updated"
    - "updatedBy"
    properties:
      created:
        type: "string"
        format: "date-time"
        description: "When the entity was created."
      updated:
        type: "string"
        format: "date-time"
        description: "When the entity was updated."
      createdBy:
        type: "string"
        format: "UUID"
        description: "Creator of the entity."
      updatedBy:
        type: "string"
        format: "UUID"
        description: "User that last updated the entity."
    description: "Describes the audit fields that are present in all the models in\
      \ this API."
...
  Skill:
    allOf:
    - type: "object"
      required:
      - "externalId"
      - "id"
      - "name"
      - "skillProviderId"
      - "uri"
      properties:
        id:
          type: "string"
          format: "UUID"
          description: "The skill id"
        skillProviderId:
          type: "string"
          format: "UUID"
          description: "The referenced skill provider id"
        name:
          type: "string"
          description: "The name of the skill"
        externalId:
          type: "string"
          description: "The external id for the skill"
        uri:
          type: "string"
          description: "The uri for the skill"
    - $ref: "#/definitions/AuditFields"
  SkillRequestBody:
    allOf:
    - type: "object"
    - $ref: "#/definitions/SkillUpdateRequestBody"
  SkillUpdateRequestBody:
    type: "object"
    properties:
      skillProviderId:
        type: "string"
        format: "UUID"
        description: "The id of provider for this Skill."
      name:
        type: "string"
        description: "Name of Skill"
      uri:
        type: "string"
        description: "Uri of Skill"
      externalId:
        type: "string"
        description: "External Id of skill"
    example:
      skillProviderId: "skillProviderId"
      name: "name"
      externalId: "externalId"
      uri: "uri"
...
  Role:
    allOf:
    - type: "object"
      required:
      - "id"
      - "name"
      properties:
        id:
          type: "string"
          format: "UUID"
          description: "The id of the role."
        name:
          type: "string"
          description: "The name of the role."
    - $ref: "#/definitions/AuditFields"
...
  NameRequestBody:
    type: "object"
    required:
    - "name"
    properties:
      name:
        type: "string"
        description: "Name of the entity"
    description: "Simple request body containing the name of the entity."
    example:
      name: "name"
...
  Unauthorized:
    type: "object"
    properties:
      message:
        type: "string"
        example: "Unable to authenticate the user."
        description: "The unauthorized error message."
    description: "The unauthorized error entity."
  NotFound:
    type: "object"
    properties:
      message:
        type: "string"
        example: "A resource with the name could not be found."
        description: "The not found error message."
    description: "The not found error entity."
  ServerError:
    type: "object"
    properties:
      message:
        type: "string"
        example: "Something went wrong while processing your request. We�re sorry\
          \ for the trouble. We�ve been notified of the error and will correct it\
          \ as soon as possible. Please try your request again in a moment."
        description: "The server error message."
    description: "The server error entity."
  BadRequest:
    type: "object"
    properties:
      message:
        type: "string"
        example: "Invalid input."
        description: "The bad request error message."
    description: "The bad request error entity."
  Forbidden:
    type: "object"
    properties:
      message:
        type: "string"
        example: "You are not allowed to access the request."
        description: "The forbidden error message."
    description: "The permission error entity."
  Conflict:
    type: "object"
    required:
    - "message"
    properties:
      message:
        type: "string"
        example: "Creating a resource with a name already exists."
        description: "The conflict error message."
    description: "The conflict error entity."